Migrate params and variables
