+++ CHANGELOG.md
@@ -19,6 +19,12 @@
 
 * Changement mineur.
 * Détails :
+  - Régorganise les paramètres en créant marche_travail
+
+### 0.34.O - [#147](https://github.com/openfisca/openfisca-tunisia/pull/147)
+
+* Changement mineur.
+* Détails :
   - Migre vers pyproject.toml
 
 ### 0.33.4 - [#135](https://github.com/openfisca/openfisca-tunisia/pull/135)
