+++ openfisca_tunisia/parameters/marche_travail/indemnite_complementaire_provisoire.yaml
@@ -7,3 +7,4 @@ metadata:
   reference:
     1981-04-07:
       title: Décret n° 81-437 du 7 avril 1981
+documentation: Le smic est composé du salaire de base et de cette indemnité complémentaire provisoire et de sa majoration
