+++ openfisca_tunisia/variables/prelevements_obligatoires/impot_revenu/revenus_categoriels/__init__.py
@@ -1,0 +1,1 @@
+'''Revenus catégoriels.'''
